se agrego mappers y se arreglo los services

+++ src/main/java/com/gym/GestorGym/service/ClaseService.java
@@ -1,6 +1,7 @@
 package com.gym.GestorGym.service;
 
 import com.gym.GestorGym.dto.ClaseDTO;
+import com.gym.GestorGym.mapper.ClaseMapper;
 import com.gym.GestorGym.models.Clase;
 import com.gym.GestorGym.repository.ClaseRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,29 +13,37 @@ import java.util.List;
 public class ClaseService {
     @Autowired
     private ClaseRepository claseRepository;
+    @Autowired
+    private ClaseMapper claseMapper;
 
     public void crear(ClaseDTO claseDTO) {
-        Clase clase = new Clase();
-        clase.setDescripcion(claseDTO.getDescripcion());
+        Clase clase = claseMapper.toEntity(claseDTO);
         claseRepository.save(clase);
     }
 
-    public List<Clase> lista(){
-        return claseRepository.findAll();
+    public List<ClaseDTO> lista(){
+        List<Clase> clases = claseRepository.findAll();
+        return claseMapper.toList((clases));
     }
 
-    public Clase listaId(int id) {
-        return claseRepository.findById(id)
+    public ClaseDTO listaId(int id) {
+        Clase clase = claseRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Clase no encontrada"));
+        return claseMapper.toDto(clase);
     }
 
-    public void update(int id, ClaseDTO claseDTO) {
-        Clase clase = listaId(id);
-        clase.setDescripcion(claseDTO.getDescripcion());
-        claseRepository.save(clase);
+    public ClaseDTO update(int id, ClaseDTO claseDTO) {
+        Clase clase = claseRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Clase no encontrada"));
+        claseMapper.updateEntity(claseDTO, clase);
+        Clase actualizado = claseRepository.save(clase);
+        return claseMapper.toDto(actualizado);
     }
 
     public void delete(int id) {
+        if (!claseRepository.existsById(id)) {
+            throw new RuntimeException("Clase no encontrada");
+        }
         claseRepository.deleteById(id);
     }
 }
